Checking

+++ src/main/java/com/driver/services/impl/UserServiceImpl.java
@@ -6,32 +6,44 @@ import com.driver.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Optional;
+
 @Service
 public class UserServiceImpl implements UserService {
 
     @Autowired
     UserRepository userRepository4;
-
     @Override
     public void deleteUser(Integer userId) {
+        Optional<User> user=userRepository4.findById(userId);
+        if(user.isEmpty())
+        {
+           throw new RuntimeException("User doesn't exist");
+        }
         userRepository4.deleteById(userId);
     }
 
     @Override
     public User updatePassword(Integer userId, String password) {
-        User user = userRepository4.findById(userId).get();
-        user.setPassword(password);
-
-        return userRepository4.save(user);
+        Optional<User> user=userRepository4.findById(userId);
+        if(user.isEmpty())
+        {
+            throw new RuntimeException("User doesn't exist");
+        }
+        User user1=user.get();
+        user1.setPassword(password);
+        userRepository4.save(user1);
+        return user1;
     }
 
     @Override
     public void register(String name, String phoneNumber, String password) {
-        User user = new User();
+        User user=new User();
         user.setName(name);
-        user.setPhoneNumber(phoneNumber);
         user.setPassword(password);
-
+        user.setPhoneNumber(phoneNumber);
+        user.setReservationList(new ArrayList<>());
         userRepository4.save(user);
     }
 }
